Extract HandlerMapping struct to seperate file
